Added Execution for JAL/JALR and start chaing 32 to 64 bits types all basic structs

+++ Run.fs
@@ -11,9 +11,9 @@ open ISA.RISCV.CLI
 open ISA.RISCV.Decode
 
 // Help function for fetch Elf data
-let getSectionContent (section : ProgBitsSection<uint32>) =
-    let fetchIndexAddr (data : byte array) (index : uint32) =
-        Array.zip [| index .. index + uint32(data.Length) - 1u |] data
+let getSectionContent (section : ProgBitsSection<int64>) =
+    let fetchIndexAddr (data : byte array) (index : int64) =
+        Array.zip [| index .. index + int64(data.Length) - 1u |] data
 
     if section.Flags.HasFlag SectionFlags.Executable then
         fetchIndexAddr (section.GetContents()) section.LoadAddress
